requirements.txt
selenium>=4.25.0
webdriver-manager>=4.0.2
beautifulsoup4>=4.12.3
pandas>=3.0.0
openpyxl>=3.1.5
PyYAML>=6.0.2
schedule>=1.2.2
requests>=2.32.3
lxml>=5.3.0
flask>=3.1.0
